Coding - Replace C-style casts with C++ cast operators
Applies clang-tidy 20.1.8 google-readability-casting across all
buildable modules (FoundationClasses, ModelingData, ModelingAlgorithms,
Visualization, DataExchange, ApplicationFramework, Draw).

751 files changed; clang-tidy replaced C-style casts with static_cast /
reinterpret_cast / const_cast or removed unnecessary ones. Four manual
corrections were needed where the tool's default choice did not
compile:

- OSD_SharedLibrary.cxx: dlsym() returns void*, converting to function
  pointer requires reinterpret_cast, not static_cast.
- TopLoc_Location.lxx: casts to `const void**` from const member
  address dropped a qualifier; rewrote as `const void* const*`.
- TopOpeBRep_sort.cxx: qsort callback casts from `const void*` needed
  `const BOA_t*` to preserve const.
- XCAFDoc_ShapeTool.cxx: casts to `void**` from const handle address
  dropped qualifiers; rewrote as `void* const*`.
- IntCurveSurface_InterUtils.pxx: tool emitted an extra closing
  parenthesis in two `(double)iU`-style casts; removed.

Excluded: src/Deprecated/, Khronos headers (OpenGl_glext.h,
OpenGl_khrplatform.h).

Verified by full Release build (Linux, GCC 15.2, Ninja, -O3 -DNDEBUG)
across all seven modules. 0 errors.

+++ src/DataExchange/TKDEPLY/RWPly/RWPly_PlyWriterContext.cxx
@@ -236,24 +236,24 @@ bool RWPly_PlyWriterContext::WriteVertex(const gp_Pnt&                    thePoi
 
   if (myIsDoublePrec)
   {
-    *myStream << (double)thePoint.X() << " " << (double)thePoint.Y() << " " << (double)thePoint.Z();
+    *myStream << thePoint.X() << " " << thePoint.Y() << " " << thePoint.Z();
   }
   else
   {
-    *myStream << (float)thePoint.X() << " " << (float)thePoint.Y() << " " << (float)thePoint.Z();
+    *myStream << static_cast<float>(thePoint.X()) << " " << static_cast<float>(thePoint.Y()) << " " << static_cast<float>(thePoint.Z());
   }
   if (myHasNormals)
   {
-    *myStream << " " << (float)theNorm.x() << " " << (float)theNorm.y() << " "
-              << (float)theNorm.z();
+    *myStream << " " << static_cast<float>(theNorm.x()) << " " << static_cast<float>(theNorm.y()) << " "
+              << static_cast<float>(theNorm.z());
   }
   if (myHasTexCoords)
   {
-    *myStream << " " << (float)theUV.x() << " " << (float)theUV.y();
+    *myStream << " " << static_cast<float>(theUV.x()) << " " << static_cast<float>(theUV.y());
   }
   if (myHasColors)
   {
-    *myStream << " " << (int)theColor.r() << " " << (int)theColor.g() << " " << (int)theColor.b();
+    *myStream << " " << static_cast<int>(theColor.r()) << " " << static_cast<int>(theColor.g()) << " " << static_cast<int>(theColor.b());
   }
   *myStream << "\n";
   if (++myNbVerts > myNbHeaderVerts)
